requirements.txt
Flask
SQLAlchemy
Flask-SQLAlchemy
asyncpg
aiohttp
Gino
Werkzeug
pydantic
python-dotenv
isort